+++ src/components/films-section.js
@@ -1,0 +1,13 @@
+const createFilmsSectionTemplate = () => {
+  return (`
+  <section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+
+      <div class="films-list__container"></div>
+    </section>
+  </section>
+  `);
+};
+
+export {createFilmsSectionTemplate};
