+++ src/components/films-sub-section.js
@@ -1,0 +1,11 @@
+const createFilmExtraListTemlate = (title) => {
+  return (`
+  <section class="films-list--extra">
+    <h2 class="films-list__title">${title}</h2>
+
+    <div class="films-list__container"></div>
+  </section
+  `);
+};
+
+export {createFilmExtraListTemlate};
